Changed boundary conditions, upgraded a_function
This was a big update. The most important thing I changed was the
boundary conditions. I hadn't attempted to implement "correct" BC
before. This time I followed advice from "Computer simulation of wave
collapses in the nonlinear Schrodinger Equation" where they recommend
that you approximate points outside of the grid with polynomial
extrapolation (formula for extrapolation is given in paper).

I changed the differencing scheme from O(drho^4) accuracy to O(drho^2).
The reason I did this was so I only have to extrapolate a single
external grid point (rather than the two points required in the O(dr^4)
scheme.

I also implemented a new "a" function. This one is based around the
magnitude of psi on-axis. I was suspicious of the original a function
because there were some arbitrary constants out in front of it from the
paper, and I wasn't sure if they effected the arbitrary constants out in
front of the L function.  After testing I am quite sure they do not.

The original a function, which I call 'gradpsi_L2' should be initialized
with L = 1. The new a function, which is called 'psi_mag' should also be
initialized with L = 1. The newer a function is less desirable because
the solution gets skinnier while propagating, and eventually gets too
skinny for the grid points. However, I understand the mathematics behind
it completely and thus it's a good check against the better, older "a"
function which I wasn't sure was implemented correctly.